Finish demo initial stage
